Turn off Style/FrozenStringLiteralComment cop
This cop is designed to help upgrade to Ruby 3.0. It requires the
comment `#frozen_string_literal: true` to the top of files to enable
frozen string literals.

Frozen string literals will be default in Ruby 3.0 but is only valid in
Ruby 2.3+. We probably want to revisit this when handling Ruby 3.0 but
since Ruby 3.0 will not be released soon, there might be more important
things to do in the meanwhile.